Catch wrong Property type when loading settings.
I'm not sure how it happened but I think a setting saved in my blender file was of the incorrect type for it's associated Property. I had to make this change to be able to export again.

It seems fairly inconspicuous to me, but I would understand if you'd refuse the patch without a test case or example blend file (as mine no longer has the culprit setting saved).