Exporter: correctly convert UV texture transforms
Previous to this change, the values in the Mapping node in Blender
were copied straight into the KHR_texture_transform object. This
is incorrect because the Mapping node is applied in Blender's UV
space, which is related to glTF's with the u,v->u,1-v conversion.
That meant that a Mapping node that translates up would be exported
as a KHR_texture_transform that translates down because of the
differing directions of the V axis.

The correct thing to do is to calculate the equivalent transform
in the other UV space.

This was added to the common module because the importer would
need to apply the same conversion to get the right Mapping node
from a KHR_texture_transform.