Import: factor out code for creating texture shader nodes
The code that creates the string of UV Map -> Mapping -> Tex Img
nodes is now in one place. Removes many bugs and inconsistencies
in the duplicated code, and nice code size reduction too.

Also removes all the existing code for storing textureInfo stuff
on texture or info objects, or setting the name of UV layers from
the mesh classes (the name was always just TEXCOORD_n anyway).
IOW, it only needs to look at the textureInfo object to create
the needed nodes.