Export: avoid pointless image->python data->image roundtrips
The glTF spec requires that eg. the metalllicRoughness texture have the
metalness in B and the roughness in G. The exporter will generate the
appropriate image for this even if the material graph reads the metalness
from, say, the R channel of one image and the roughness from the G channel
of another.

To do this, *all* images are decoded to python data. The code sequence for
this looks roughly like this:

    creating export image:
        pixel_data = np.array(blender_image.pixels[:])           (★1)

    pixel data may be modified to insert channels

    encoding export image:
        if created from blender image that's on disk:
            read back data from the disk

        tmp_image = bpy.data.images.new(...)
        tmp_image.pixels = pixel_data.img.flatten().tolist()     (★2)
        save tmp_image to disk and read it back

The lines marked with a ★ are slow and eat up loooots of memory. Eg.
blender_image.pixels[:] is going to allocate a tuple of width*height*channel
floats. It's really bad.

As you can see, sometimes it was possible to skip ★2 but everything did ★1.

However, for many images, like all color textures, normals, or
metal/rough/occlusion textures that are already packed the way the exporter
likes, this is totally wasted effort: the image data is converted to python
data, nothing gets changed, and the python data is converted back to a new
tmp image, which is then encoded to disk and read back.

In this case it would have been better to just encode the original image!

So here's the solution:

Change ExportImage to store some declarative data saying how it's channels
should be filled in (ex. it says its R channel should be filled by the G channel
of such-and-such Blender image) instead of storing a numpy array.

Then when it comes time to encode, we examine that data and determine
whether all we need to do is save some existing Blender image (happy path),
or whether we need to do the pixel manipulation to repack the channels
(unhappy path).

The unhappy path is basically the same as before, but the happy path is
significantly faster and uses much less memory :)