Export: use image.save_render to save images
It's strangely difficult to figure out how to save a copy of an
image in Python

The UI for "Save a Copy" uses bpy.ops.image.save_as(copy=True) but
I couldn't figure out how to use that from a script.

So bpy.types.Image has two methods for saving: save() and save_render().
save() saves to the image's filepath field. save_render() takes the
path to save to and uses the scene's render settings.

The previous code for this would overwrite the filepath and format
of the image, call image.save(), and then restore the original
values. But this had the problem of also clearing the dirty flag
on dirty images (since they did get saved after all) and of turning
generated images into file-on-disk images.

So I'm abandoning that idea and using the other one, save_render(),
since that doesn't modify the image. Unfortunately for us it picks
up settings from the scene, but since it takes a scene argument,
we can create a temp scene and use that.

What settings do we need to set? Looking at the Blender source code
for rna_Image_save_render, it uses

    scene->view_settings
    scene->display_settings
    scene->r.im_format
    scene->r.im_format.planes
    scene->r.dither_intensity

Setting the display_device to 'None' should take care of
display_settings (and testing shows that 'None' gives the right
output regardless of the image's colorspace).

Since the temp scene is newly created, the default view settings
(eg. exposure = 1) should all be okay.

im_format is obviously the format to save (eg. PNG).

Planes should be the number of bitplanes (ie. 24 for 8-bit RGB) I
think. Probably don't need to worry about that.

Dithering doesn't appear to do anything as far as I can tell, but set
it to zero anyway just in case.

Crossing my fingers that this is right this time.