Export: avoid pointless image<->Python data conversions
Previously all images were decoded to a Python tuple of 4*w*h floats
and put into a numpy array. This was so the channels could be
assembled together for eg. metallicRoughness textures. After assembly,
the numpy array would then be converted back to a tuple, fed into the
image.pixels of a new Blender image, the image saved, and the encoded
image read back.

A special path handled images that were already saved on disk;
the file on disk would just be read, skipping the final
array->tuple->image conversion, but the initial conversion
image->tuple->array was always performed.

For most images this is wasted time and memory. The channels aren't
going to be changed, so we could have just used the original image!

So this commit makes ExportImage smarter. It now stores a description
of how its intended image should be built instead of an array of
pixel values. For example, it says its R channel should be filled by
the G channel of such-and-such image.

Then when we go to encode, we have enough information to make better
decisions about how to do it. There is a "happy path" where we
can just use an existing Blender image and don't ever have to
convert to Python data. The "unhappy" path, where we do have to
assemble an image out of channel, however, remains the same.

The happy path should contain all color image (baseColor, emission),
normal maps, and metal-rough-occlusion textures that are already
packed the way the exporter likes.

Knocks about 50% off the export time for models with lots of texture
data, and reduces memory usage.