Import: rewrite material code
Rewrites material code as functions which each make one "piece" of
the node tree, taking the location to create at and the sockets to
attach to as arguments. Results in better composability.

Several bugs were fixed. Notably the base color term is now
calculated correctly, nodes are layed out properly, and the
pbrSpecularGlossiness setup is improved.