Export: add option to export textures into a folder
For the "glTF Separate" format, you can now pick a path relative
to the location of the .gltf to put texture files in. Defaults to
the empty path, ie. the same folder as the .gltf, so there's no
change in the default behaviour.

Since the __add_image function needed to know the output path,
this takes the suggestion from the TODO there and passes the
export options to the constructor of the Exporter object.