Import: slightly simplify vert skinning code
Pre-calculate the joint matrices so we only have to do one matrix-
vector multiply for each term instead of two.

Also compute the weight sums during the main loop and divide at the
end.

This is very slightly faster, but actually skinning the verts
doesn't seem to take that long to begin with.