Import: remove "husks" left behind when moving skinned meshes
When a skinned mesh is moved to a child of its arma, the old node
where it was may be left with no purpose, becoming an empty "husk".
When it won't mess up anything (animations, other children, etc)
we can move the entire node to become a child of the arma instead.

One very important consequence: when round-tripping, even if a
skinned mesh does get moved to a new node on the *first* import,
on all subsequent imports, it can stay at that node, with no
further changes to the node tree.