Export: don't combine image names if they're all from the same file
Fix #1014

Before this, if there was one image in an ExportImage, we'd try to
use its filename as the image name. This generalizes this to when
there's one filepath (ie. multiple images with the same filepath).

The __gather_name function was also changed to examine the ExportImage
rather than traverse the node graph, which is simpler.