Export: handle different-sized images in ExportImage
When combining image channels, scale all images to the output
image dimensions before combining.

The output image width is the max of all input image widths (and
similarly for the height), so images are always scaled up, never
down (which would be lossy).

Probably still better to have them all be the same size though
(as far as efficiency)...