Export/Perf: accept numpy arrays for primitive attribute arrays
Accept primitive arrays in the form of numpy arrays in
addition to lists (not currently used but extract_primitives
can be enhanced for it).

Also convert lists to numpy arrays instead of array.arrays to
use tobytes on.

Also calculates the min/max with numpy for position attributes.

This is a bit faster even without enhancing extract_primitives.