Export: stopgap for zero-weight verts producing invalid files
Verts without any vertex group assignments currently export
with zero weights, which is invalid and renders differently in
different viewers.

This is a stopgap that produces a valid file. It does the same
thing the Three.js viewer does; treat these verts as assigned
to the 0th joint.

These verts SHOULD act as if they weren't skinned and this
commit does NOT FIX that (unless the root bone happens to have
a default transform, so that verts skinned by it happen to
act as if they weren't skinned).