Export: find factors when there are textures
Adds get_factor from socket to read a factor from a socket directly
or from a MULTIPLY node in front of the socket.

This is used for baseColorFactor, roughnessFactor, etc. The test
from the last commit now passes.

The RGB part of baseColorFactor already had code for this that
worked differently (it didn't require MULTIPLY be right in front
of the socket and it required there be a texture). A log statement
was also dropped.s