Export: detect when a file will be overwritten
In the file select dialogue for the exporter, the filename should
be red when there's an existing file that will be overwritten.

Before this we only used the stem of the filename (ie a not a.glb)
so it wouldn't detect that a.glb would be overwritten.

This changes it to use the full filename with the extension. The
filename has to be updated when either the name or format option
changes.