Export: correctly handle empty/OOB material slots
If material_index >= number of material slots, the old code would
export this as material: None. This fixes it to use the last material
slot, which is what Blender does.

Also because slots without a material were skipped in the material_names
list, this messed up material indices. Now empty material slots export
with material: None.