Workaround to get shape key normals faster
Getting shape key normals with ShapeKey.normals_split_get() is
particularly slow because Blender first has to iterate the data into a
tuple and then NumPy has to iterate that tuple into an array.

To work around these iterations, the shape key coordinates can be copied
into a mesh's 'position' attribute and then the normals can be
retrieved from the mesh's `corner_normals` property after forcing the
mesh's normals to be recalculated.

This makes use of the new ShapeKey.points property added in Blender
4.1.0 that has faster access through `foreach_get` than ShapeKey.data.

For large meshes and/or meshes with many shape keys, this can reduce
export times by a few seconds.