Remove setting the read-only 'loop_total' property
'loop_total' is a read-only property since Blender 3.6.

Since Blender 3.6, the number of loops/corners of each face is derived
by subtracting the 'loop_start' of the current face from the
'loop_start' of the next face (internally there is an extra 'loop_start'
at the end for the last face to use).