Update edge vertices import for Blender 3.6
In Blender 3.6, the '.edge_verts' attribute replaces `MeshEdge.vertices`
accessed through `Mesh.edges`.

The old access still exists, and, unlike most access that has been
replaced with attributes, the old access is the same speed as the newer
access due to an implementation detail.

No performance difference is expected with this patch.