Describe mesh_node_idx more clearly
The previous comment made it unclear whether `mesh_node_idx`
refers to the index in the `meshes` array or the `nodes` array.

Further, the use of the word "Indices" implies that multiple,
different index values are present for a single VNode.
As best as I can tell, two of the three attributes will always be `None`,
with only one attribute getting the index in the "nodes" collection,
depending on the type of object (camera, light, or "mesh" node).