Allow Extensions to pass data through without __fixing
Compared to extras, extensions currently don't allow bypassing the __fix_json function, which results in extras being able to export values with empty objects, but extensions can not.
Here is an example of the same data exported through extras and extensions.
Note how the extension data is empty.

```json
"nodes":[
    {
        "extensions":{
            "EXT_my_extension":{
                "components":[
                    {}
                ]
            }
        },
        "extras":{
            "components":[
                {
                    "test_components::Marker":{}
                }
            ]
        },
        "mesh":0,
        "name":"Cube"
    }
],
```

This PR enables export extensions to define which of their glTF extensions should be passed through without using `__fix_json`.
The result is that the keys with empty dict values are kept in the same way extras keeps them.

```
"nodes":[
    {
        "extensions":{
            "EXT_my_extension":{
                "components":[
                    {
                        "test_components::Marker":{}
                    }
                ]
            }
        },
        "extras":{
            "components":[
                {
                    "test_components::Marker":{}
                }
            ]
        },
        "mesh":0,
        "name":"Cube"
    }
],
```